beta 4

+++ FortniteLauncher/Server/DeleteServer.cs
@@ -13,7 +13,18 @@ namespace FortniteLauncher.Server
     {
         public void DeleteSRV()
         {
+            if (!Directory.Exists("LauncherData/LawinServer/LawinServer-main"))
+            {
+                ContentDialog dialog = new ContentDialog();
+                dialog.Title = "Error";
+                dialog.Content = "LawinServer is not installed! Try clicking Install button";
+                dialog.CloseButtonText = "OK";
+                dialog.ShowAsync();
+            }
+            else if (Directory.Exists("LauncherData/LawinServer/LawinServer-main"))
+            {
                 Directory.Delete("LauncherData/LawinServer/LawinServer-main", true);
+            }
         }
     }
 }
